Post now posts to a specified table
You can now use /post to post data to any table. The json should have
the field 'table' with the table name, and the field 'content' with the
actual data you want to insert into the table.
Made the token creation work via this new post method too, and deleted
the tokenreceiver function (as it is now redundant).
Also some further cleaning up of the code.