Check if table exists for post + made specific post for creating tokens + node modules
I added a check to see if the table you want to insert data into
actually exists. So now it doesn't crash if it tries to insert into a
table that doesn't exist, it returns an error message.
I also made a specific post for creating tokens, so I can add the token
check to the regular post.
There was a problem with some node modules only being available for the
data generator and not for the api, so I made sure they both have the
same modules available.